--- a/package.json
+++ b/package.json
@@ -111,7 +111,6 @@
     "babel-preset-react": "^6.24.1",
     "babel-register": "^6.26.0",
     "babel-runtime": "^6.26.0",
-    "base64-loader": "^1.0.0",
     "copy-webpack-plugin": "^4.5.1",
     "cors": "^2.8.4",
     "cross-env": "^5.1.4",