This is a **work in progress**, mostly complete just requires some testing

 - Use the browsers fetch rather than the request module
 - base64-loader -> raw-loader
 - Remove ol3 for now because it's been broken for a while
 - Removed old GitHub gist support as it's no longer functional
 - Removed Mousetrap as we were only using a small part of the functionality
 - Moved to single js file to make things simplier

Here are some stats:
```
               vendor (before)   app (before)   vendor (after)   app (after)
Stat size:     3.26MB            4.5MB          n/a              3.42MB
Parsed size:   1.95MB            1.9MB          n/a              2.09MB
Gzipped size:  549.64KB          449.93KB       n/a              507.41KB
```
```
Gzipped total size before: 549.64KB + 449.93KB  = 999.57KB
Gzipped total size after: 507.41KB (saving ~49% saving)
```

```
Parsed total size before: 1.95MB + 1.9MB  = 3.85MB
Parsed total size after: 2.09MB (saving ~45% saving)
```

This should also improve startup time also. Local js parse times

Before

```
vendor.js 267.46ms
app.js    242.23ms
total     509.69ms
```

After

```
vendor.js  n/a
app.js     341.22ms (~33% speed up)
```

Full results

 - Before: https://343-67726921-gh.circle-artifacts.com/0/artifacts/build/bundle-stats.html
 - After: https://745-84182601-gh.circle-artifacts.com/0/artifacts/build/bundle-stats.html

Note: that the removed OpenLayers dependency accounted for a major chunk of those savings when we fix and re-enable OpenLayers we should make sure it's loaded externally from main bundle and only loaded when required.